Added validation step
also:
Added variable scope to generator so that it's parameters can be shared with other hypernetwork instances.
In batch norm, changed to a manual moving average, since otherwise the parameters are not shared in scope.
Moved ResnetWeights() to a separate file.
Changed ResnetWeights() to be a dictionary so that sess.run() can fetch it.